--- models/user.js
@@ -1,17 +1,0 @@
-const mongoose = require('mongoose')
-const { schemaOptions } = require('./modelOptions')
-
-const userSchema = new mongoose.Schema({
-  username: {
-    type: String,
-    required: true,
-    unique: true
-  },
-  password: {
-    type: String,
-    required: true,
-    select: false
-  }
-}, schemaOptions)
-
-module.exports = mongoose.model('User', userSchema)
