+++ models/userModel.js
@@ -1,77 +1,17 @@
-const mongoose = require("mongoose");
-const bcrypt = require("bcrypt");
-const validator = require("validator");
+const mongoose = require('mongoose')
+const { schemaOptions } = require('./modelOptions')
 
-const Schema = mongoose.Schema;
-
-const userSchema = new Schema({
-  email: {
+const userSchema = new mongoose.Schema({
+  username: {
     type: String,
     required: true,
-    unique: true,
+    unique: true
   },
   password: {
     type: String,
     required: true,
-  },
-
-  occupation: {
-    type: String,
-    required: true,
-  },
-  firstName: { type: String },
-  lastName: { type: String },
-  student: Boolean,
-  professional: Boolean,
-});
-
-// static signup method
-userSchema.statics.signup = async function (email, password, occupation) {
-  // validation
-
-  if (!email || !password) {
-    throw Error("All fields must be filled");
+    select: false
   }
-  if (!validator.isEmail(email)) {
-    throw Error("Email not valid");
-  }
+}, schemaOptions)
 
-  const exists = await this.findOne({ email });
-
-  if (exists) {
-    throw Error("Email already in use");
-  }
-
-  const salt = await bcrypt.genSalt(10);
-  const hash = await bcrypt.hash(password, salt);
-
-  const user = await this.create({
-    email,
-    password: hash,
-    occupation,
-  });
-
-  return user;
-};
-
-// static login method
-userSchema.statics.login = async function (email, password, occupation) {
-  if (!email || !password) {
-    throw Error("All fields must be filled");
-  }
-
-  const user = await this.findOne({ email });
-  if (!user) {
-    throw Error("Incorrect email");
-  }
-
-  const match = await bcrypt.compare(password, user.password);
-  if (!match) {
-    throw Error("Incorrect password");
-  }
-
-  return user;
-};
-
-const User = mongoose.model("User", userSchema);
-module.exports = User;
+module.exports = mongoose.model('User', userSchema)
