+++ package.json
@@ -2,10 +2,13 @@
   "dependencies": {
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
+    "crypto-js": "^4.1.1",
     "dotenv": "^16.0.3",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
+    "express-validator": "^6.15.0",
     "gridfs-stream": "^1.1.1",
     "helmet": "^6.0.1",
     "jsonwebtoken": "^9.0.0",
@@ -18,10 +21,10 @@
   },
   "name": "sync-backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "nodemon app.js"
   },
   "keywords": [],
   "author": "",
