+++ routes/auth.js
@@ -1,9 +1,9 @@
 const router = require('express').Router()
-const userController = require('../controllers/user')
+const userController = require('../controllers/userController')
 const { body } = require('express-validator')
 const validation = require('../handlers/validation')
 const tokenHandler = require('../handlers/tokenHandler')
-const User = require('../models/user')
+const User = require('../models/userModel')
 
 router.post(
   '/signup',
