+++ routes/project.js
@@ -2,7 +2,7 @@ const router = require('express').Router()
 const { param } = require('express-validator')
 const validation = require('../handlers/validation')
 const tokenHandler = require('../handlers/tokenHandler')
-const projectController = require('../controllers/project')
+const projectController = require('../controllers/projectController')
 
 router.post(
   '/',
