--- views/editProjectsKanban.ejs
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Edit Project Kanban Page</title>
-  </head>
-  <body>
-    <h1>Edit Project Kanban Page</h1>
-    <form action="/projects/<%=kanban.id%>/kanban?_method=PUT" method="POST">
-      Kanban ID: <%=kanban.id%>
-      <br /><br />
-      To-do:<br />
-      <input
-        type="text"
-        name="todo"
-        placeholder="<%=kanban.todo%>"
-      /><br /><br />
-      Doing:<br />
-      <input
-        type="text"
-        name="doing"
-        placeholder="<%=kanban.doing%>"
-      /><br /><br />
-      Done:<br />
-      <input
-        type="text"
-        name="done"
-        placeholder="<%=kanban.done%>"
-      /><br /><br />
-      <input type="submit" name="" value="SUBMIT" />
-    </form>
-    <br />
-    <nav>
-      <a href="/projects/<%=kanban.id%>">Back to Project</a>
-    </nav>
-  </body>
-</html>
