--- views/main.ejs
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <title>Main Page</title>
-  </head>
-  <body>
-    <h1>Main Page</h1>
-    <a href="/users">Users Index Page</a><br /><br />
-    <a href="/projects">Projects Index Page</a>
-  </body>
-</html>
