--- views/newProjects.ejs
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>New Project</title>
-  </head>
-  <body>
-    <h1>New Project</h1>
-    <form action="/projects" method="POST">
-      Project Name: <input type="text" name="projectName" /><br /><br />
-      Agile: <input type="checkbox" name="agile" /><br /><br />
-      Kanban: <input type="checkbox" name="kanban" /><br /><br />
-      <input type="submit" name="" value="CREATE PROJECT" />
-    </form>
-    <br />
-    <nav>
-      <a href="/projects">Back to Projects Index</a>
-    </nav>
-  </body>
-</html>
