--- views/newUsers.ejs
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>New User</title>
-  </head>
-  <body>
-    <h1>New User</h1>
-    <form action="/users" method="POST">
-      Email: <input type="text" name="email" /><br /><br />
-      Password: <input type="text" name="password" /><br /><br />
-      First Name: <input type="text" name="firstName" /><br /><br />
-      Last Name: <input type="text" name="lastName" /><br /><br />
-      Student: <input type="checkbox" name="student" /><br /><br />
-      Professional: <input type="checkbox" name="professional" /><br /><br />
-      <input type="submit" name="" value="CREATE USER" />
-    </form>
-    <br />
-    <nav>
-      <a href="/users">Back to Users Index</a>
-    </nav>
-  </body>
-</html>
