--- views/showProjects.ejs
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Projects Show Page</title>
-  </head>
-  <body>
-    <h1>Projects Show Page</h1>
-    Project Name: <%=project.projectName; %><br /><br />
-    Agile: <% if (project.agile === true) { %> Yes<br /><br />
-    <% } else { %> Not Applicable.<br /><br /><% } %> Kanban: <% if
-    (project.kanban === true) { %> Yes<br /><br />
-    <% } else { %> Not Applicable.<br /><br /><% } %> <% if
-    (!kanban.tasksAssigned) { %> Tasks: Pending Assignment <% } else { %>
-    Tasks:<br />
-    To-do: <%=kanban.todo%><br />
-    Doing: <%=kanban.doing%><br />
-    Done: <%=kanban.done%><br />
-    <% } %>
-    <br /><br />
-    <nav>
-      <a href="/projects">Back to Projects Index</a>
-    </nav>
-  </body>
-</html>
