--- views/showUsers.ejs
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Users Show Page</title>
-  </head>
-  <body>
-    <h1>User Details</h1>
-    First Name: <%=user.firstName; %><br />
-    Last Name: <%=user.lastName; %><br />
-    Student: <% if(user.student === true){ %>Yes<br />
-    <% } else { %>Not Applicable.<br /><br />
-    <% } %> Professional: <% if(user.professional === true){ %>Yes<br /><br />
-    <% } else { %>Not Applicable.<br /><br />
-    <% } %>
-    <nav>
-      <a href="/users">Back to Users Index</a>
-    </nav>
-  </body>
-</html>
